requirements-web.txt
Flask==2.3.3
tensorflow==2.20.0
keras==3.11.3
numpy==2.1.3
opencv-python==4.10.0.84
Pillow==11.3.0
h5py==3.14.0

